feat: Refactor error handling in conversation and user controllers; replace errorResponse with errorHandler and update middleware imports
